Deprecate `Modifier.onExternalDrag` and move `DragData` into ui.draganddrop
