+++ compose/ui/ui/api/desktop/ui.api
@@ -606,8 +606,24 @@ public abstract interface class androidx/compose/ui/draganddrop/DragAndDropTrans
 
 public final class androidx/compose/ui/draganddrop/DragAndDrop_desktopKt {
 	public static final fun DragAndDropTransferable (Ljava/awt/datatransfer/Transferable;)Landroidx/compose/ui/draganddrop/DragAndDropTransferable;
-	public static final fun dragData (Landroidx/compose/ui/draganddrop/DragAndDropEvent;)Landroidx/compose/ui/DragData;
+	public static final fun dragData (Landroidx/compose/ui/draganddrop/DragAndDropEvent;)Landroidx/compose/ui/draganddrop/DragData;
 	public static final fun getAwtTransferable (Landroidx/compose/ui/draganddrop/DragAndDropEvent;)Ljava/awt/datatransfer/Transferable;
+}
+
+public abstract interface class androidx/compose/ui/draganddrop/DragData {
+}
+
+public abstract interface class androidx/compose/ui/draganddrop/DragData$FilesList : androidx/compose/ui/draganddrop/DragData {
+	public abstract fun readFiles ()Ljava/util/List;
+}
+
+public abstract interface class androidx/compose/ui/draganddrop/DragData$Image : androidx/compose/ui/draganddrop/DragData {
+	public abstract fun readImage ()Landroidx/compose/ui/graphics/painter/Painter;
+}
+
+public abstract interface class androidx/compose/ui/draganddrop/DragData$Text : androidx/compose/ui/draganddrop/DragData {
+	public abstract fun getBestMimeType ()Ljava/lang/String;
+	public abstract fun readText ()Ljava/lang/String;
 }
 
 public final class androidx/compose/ui/draw/AlphaKt {
