+++ compose/ui/ui/src/desktopMain/kotlin/androidx/compose/ui/draganddrop/AwtDragData.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The Android Open Source Project
+ * Copyright 2024 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package androidx.compose.ui
+package androidx.compose.ui.draganddrop
 
+import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.graphics.toPainter
 import java.awt.Image
@@ -44,20 +45,22 @@ internal fun Transferable.dragData(): DragData {
 @OptIn(ExperimentalComposeUiApi::class)
 private object UnknownDragData : DragData
 
+@Suppress("DEPRECATION_ERROR")
 @OptIn(ExperimentalComposeUiApi::class)
 private class DragDataFilesListImpl(
     private val transferable: Transferable
-) : DragData.FilesList {
+) : DragData.FilesList, androidx.compose.ui.DragData.FilesList {
     override fun readFiles(): List<String> {
         val files = transferable.getTransferData(DataFlavor.javaFileListFlavor) as List<*>
         return files.filterIsInstance<File>().map { it.toURI().toString() }
     }
 }
 
+@Suppress("DEPRECATION_ERROR")
 @OptIn(ExperimentalComposeUiApi::class)
 private class DragDataImageImpl(
     private val transferable: Transferable
-) : DragData.Image {
+) : DragData.Image, androidx.compose.ui.DragData.Image {
     override fun readImage(): Painter {
         return (transferable.getTransferData(DataFlavor.imageFlavor) as Image).painter()
     }
@@ -80,11 +83,12 @@ private class DragDataImageImpl(
     }
 }
 
+@Suppress("DEPRECATION_ERROR")
 @OptIn(ExperimentalComposeUiApi::class)
 private class DragDataTextImpl(
     private val bestTextFlavor: DataFlavor,
     private val transferable: Transferable
-) : DragData.Text {
+) : DragData.Text, androidx.compose.ui.DragData.Text {
     override val bestMimeType: String = bestTextFlavor.mimeType
 
     override fun readText(): String {
@@ -92,3 +96,10 @@ private class DragDataTextImpl(
         return reader.readText()
     }
 }
+
+/**
+ * Returns the same value as [Transferable.dragData] but as an instance of the deprecated
+ * `androidx.compose.ui.DragData`. Remove this, along with the rest of the deprecated DnD API in 1.8
+ */
+@Suppress("DEPRECATION_ERROR")
+internal fun Transferable.deprecatedDragData() = dragData() as androidx.compose.ui.DragData
